+++ src/main/kotlin/com/alesharik/digitalgrid/utils/Recipe.kt
@@ -3,9 +3,15 @@ package com.alesharik.digitalgrid.utils
 import net.minecraft.data.recipes.RecipeCategory
 import net.minecraft.data.recipes.RecipeOutput
 import net.minecraft.data.recipes.ShapedRecipeBuilder
+import net.minecraft.data.recipes.ShapelessRecipeBuilder
 import net.minecraft.world.level.ItemLike
 
 inline fun RecipeOutput.shaped(item: ItemLike, count: Int, crossinline f: ShapedRecipeBuilder.() -> ShapedRecipeBuilder) {
     val b = ShapedRecipeBuilder.shaped(RecipeCategory.MISC, item, count)
+    f(b).save(this)
+}
+
+inline fun RecipeOutput.shapeless(item: ItemLike, count: Int, crossinline f: ShapelessRecipeBuilder.() -> ShapelessRecipeBuilder) {
+    val b = ShapelessRecipeBuilder.shapeless(RecipeCategory.MISC, item, count)
     f(b).save(this)
 }