+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 - add PLC components - watchdog and modems
 - limit PLC storage size to 64 KiB
 - make PLC API global, instead of creating peripheral
+- add casing and digibus casing + recipes